Pod::Escapes::Escapes.pm: Fill %Name2character better for EBCDIC
This commit does several things for filling in %Name2character.

1) Several special values were repeated in the file, like "&lt" meaning
"<".  Now there is only one occurrence each.

2) On Perls that have utf8::unicode_to_native (>= v5.7.3) that function
is used to get complete EBCDIC coverage, instead of the previous just a
few things.

3) On these later Perls, there is now only one code path that is used
for both ASCII and EBCDIC.  This decreases EBCDIC maintenance cost.

4) On earlier Perls, there is better EBCDIC coverage, from using
%Latin1Code_to_fallback.

The extra calls to utf8::unicode_to_native are done only at
initialization, and it is planned to have the compiler optimize these
out in v5.22.