Pod::Escapes::Escapes.pm: Generalize for EBCDIC
Some special handling for EBCDIC is unnecessary.  %Name2character
now already contains the info we need for any platform, so it can be
used where applicable and the previous special handling removed.

But numeric escapes are always Unicode, and need to be translated to
native values before finding their chr.  This commit adds a
function call to do that translation.  This is done even on ASCII
platforms where the function just returns its argument.  I did this
because it generally is better to not have separate code paths for
different platforms.  It is planned to change the Perl compiler to
optimize out this function call on ASCII platforms in Perls starting in
v5.22, so that there will be no performance hit at all for them.  And I
suspect that numeric escapes are relatively infrequently used.  But if
you would rather avoid the performance hit in earlier Perls, the code
could be refactored slightly to accomplish this, with the small extra
maintenance cost of having platform-dependent code branches.